+++ Source/dom/assets/index.js
@@ -1,0 +1,2 @@
+export { ColorAsset } from './ColorAsset'
+export { GradientAsset } from './GradientAsset'
