+++ Source/dom/index.js
@@ -10,7 +10,7 @@ export { Document, getDocuments, getSelectedDocument } from './models/Document'
 export { Library, getLibraries } from './models/Library'
 export { SharedStyle } from './models/SharedStyle'
 export { Rectangle } from './models/Rectangle'
-export { ColorAsset, GradientAsset } from './models/Asset'
+export { ColorAsset, GradientAsset } from './assets'
 
 export { Style } from './style/Style'
